# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Chess Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  email: string;
  coins: number;
  lives: number;
  unlockedLevel: number;
  dailyBonusDay: number;
  /** @nullable */
  lastDailyBonusDate: string | null;
  /** @nullable */
  lastWithdrawalDate?: string | null;
  isAdmin: boolean;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface DailyBonusInput {
  doubled: boolean;
}

export interface CoinUpdateResponse {
  coins: number;
  coinsAdded: number;
  message?: string;
}

export interface CampaignLevelInput {
  level: number;
}

export interface LevelCompleteResponse {
  coins: number;
  coinsAdded: number;
  unlockedLevel: number;
}

export interface LivesResponse {
  lives: number;
  message?: string;
}

export interface UndoResponse {
  allowed: boolean;
}

export type WithdrawalInputAmount = typeof WithdrawalInputAmount[keyof typeof WithdrawalInputAmount];


export const WithdrawalInputAmount = {
  NUMBER_10: 10,
  NUMBER_20: 20,
  NUMBER_50: 50,
} as const;

export interface WithdrawalInput {
  amount: WithdrawalInputAmount;
  /** @minLength 3 */
  upiId: string;
}

export interface WithdrawalResponse {
  id: number;
  amount: number;
  status: string;
  coins?: number;
}

export interface Withdrawal {
  id: number;
  amount: number;
  upiId: string;
  status: string;
  createdAt: string;
  coinsDeducted?: number;
}

export interface AdminWithdrawal {
  id: number;
  userId?: number;
  userEmail: string;
  amount: number;
  upiId: string;
  status: string;
  coinsDeducted: number;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalWithdrawals: number;
  pendingWithdrawals: number;
  totalCoinsInSystem: number;
}