# lib/api-zod/src/generated/types/adminWithdrawal.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Chess Game API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminWithdrawal {
  id: number;
  userId?: number;
  userEmail: string;
  amount: number;
  upiId: string;
  status: string;
  coinsDeducted: number;
  createdAt: string;
}