# lib/api-zod/src/generated/types/coinUpdateResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Chess Game API
 * OpenAPI spec version: 0.1.0
 */

export interface CoinUpdateResponse {
  coins: number;
  coinsAdded: number;
  message?: string;
}