# lib/api-zod/src/generated/types/levelCompleteResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Chess Game API
 * OpenAPI spec version: 0.1.0
 */

export interface LevelCompleteResponse {
  coins: number;
  coinsAdded: number;
  unlockedLevel: number;
}