# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Chess Game API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
}